+++ R/utils.R
@@ -825,11 +825,3 @@ make.unique.2 <- function(x, sep = ".") {
     }
   })
 }
-
-
-
-
-
-
-
-